fix/move components from pages
